fix: change validate to middleware
