+++ node-js/src/api/routes.js
@@ -3,8 +3,10 @@ import express from 'express';
 import rescue from 'express-rescue';
 import LoginController from '../controllers/login/login-controller.js';
 import ReportController from '../controllers/report/report-controller.js';
+import JWTMiddleware from '../middleware/jwt-middleware.js';
 
 const apiRoutes = express.Router();
+const jwtMiddleware = new JWTMiddleware();
 
 apiRoutes.get("/", (_req, res ) => {
   res.sendStatus(200);
@@ -12,6 +14,6 @@ apiRoutes.get("/", (_req, res ) => {
 
 apiRoutes.post('/enroll', rescue(EnrollController.postEnroll));
 apiRoutes.post('/login', rescue(LoginController.doLogin));
-apiRoutes.post('/report/all', rescue(ReportController.getAll));
+apiRoutes.post('/report/all', rescue(jwtMiddleware.validateToken), rescue(ReportController.getAll));
 
 export default { apiRoutes }
